[Mintscan API] - separate db queries for validator.go

+++ mintscan/api/models/validator.go
@@ -8,7 +8,7 @@ import (
 
 type Uptime struct {
 	Address      string `json:"address"`
-	MissedBlocks int64  `json:"missed_blocks"`
+	MissedBlocks int    `json:"missed_blocks"`
 	OverBlocks   int64  `json:"over_blocks"`
 }
